Reload list of boxes on reload.
Previously, only the status of the existing boxes would be refreshed.
That meant that adding/removing a box config required a restart. Now
it's sufficient to reload, and the list of boxes will be reloaded from
the config directory.

This also means that it's no longer an error to start up with an empty
list of boxes. After all, they can be loaded later.